About us page refactoring 🔨🧨
