--- docs/Contributing.mdx
@@ -1,5 +1,0 @@
----
-name: Contributing
----
-
-# Add Contribution/donation here!
