--- docs/Local_Installation.mdx
@@ -1,6 +1,0 @@
----
-name: Installation
----
-
-# Add Local Installation part here !
-
