--- docs/getting_started.mdx
@@ -1,13 +1,0 @@
----
-name: Getting Started
----
-
-# Hello People
-
-
-### Adipisicing officia amet exercitation minim excepteur aute sed elit qui ea.
-Dolor duis ullamco fugiat ullamco exercitation ad ad voluptate pariatur ut ex cupidatat aliquip elit culpa veniam dolore culpa.
-Lorem ipsum ut aute sit dolore mollit sed aliquip sit occaecat veniam qui proident nulla. Dolore cupidatat nisi veniam est consectetur et ut incididunt eiusmod duis.
-Anim sit fugiat exercitation duis enim enim do proident dolor ullamco sunt aute ex ut amet ut minim aute. Lorem ipsum aute culpa sunt in proident ad reprehenderit adipisicing mollit tempor consequat.
-Aliquip minim incididunt enim dolore sed sint deserunt incididunt amet est exercitation incididunt irure consequat nulla ea irure. Consectetur non dolor in cillum non reprehenderit velit ea. Aliquip nisi deserunt do in in nisi sunt consequat dolore ut velit labore. Aliqua ullamco adipisicing cillum esse tempor ad tempor incididunt elit ut. Ex amet laborum duis do sed velit ex adipisicing ut. Anim proident anim in esse quis ut nulla enim officia sed enim. Consectetur labore eu aliquip ea non in mollit in consequat dolor excepteur incididunt. Irure occaecat irure irure minim laborum dolor anim ad ut sunt aliquip. Amet ea consectetur mollit aliqua reprehenderit nisi consectetur duis magna fugiat veniam veniam. Proident voluptate ex consectetur laboris amet est non pariatur sint adipisicing exercitation. Ut ad laborum et nostrud sed magna ut id ullamco minim ut ex dolore reprehenderit. Veniam commodo magna fugiat cupidatat non voluptate cillum labore ea enim ex. Ullamco sit dolor cupidatat velit in cillum deserunt reprehenderit. Enim ut nulla dolor velit esse velit reprehenderit commodo aliquip laborum mollit pariatur est. Sint culpa labore deserunt et consequat in irure eiusmod sunt enim in quis minim. Laborum aliqua quis nulla id do aliquip et aliqua elit et aliqua exercitation eiusmod laboris laborum id dolor.
-
