+++ src/ignitus-AboutPage/ignitus-About/actionTypes.ts
