+++ src/ignitus-AboutPage/ignitus-About/reducers.ts
