+++ src/ignitus-AboutPage/ignitus-CoreTeam/Components/index.ts
