+++ src/ignitus-AboutPage/ignitus-CoreTeam/Components/team.tsx
@@ -3,9 +3,9 @@
 /* eslint-disable jsx-a11y/mouse-events-have-key-events */
 import React from 'react';
 import {Data} from '../constants';
-import {withErrorBoundary} from '../../ignitus-Shared/Components/errorBoundary';
+import {withErrorBoundary} from '../../../ignitus-Shared/Components/errorBoundary';
 import * as S from '../Styles';
-import * as T from '../../ignitus-Shared/Components/emotionStyles/shared';
+import * as T from '../../../ignitus-Shared/Components/emotionStyles/shared';
 import {coreTeam, TeamItemProps} from '../types';
 
 const PureCoreTeam = ({team}: any) => (
