+++ src/ignitus-AboutPage/ignitus-CoreTeam/index.ts
