+++ src/ignitus-AboutPage/ignitus-CoreTeam/reducers.ts
