+++ src/ignitus-AboutPage/ignitus-Team/Components/Team.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable jsx-a11y/accessible-emoji */
 import React, {useEffect} from 'react';
-import {withErrorBoundary} from '../../ignitus-Shared/Components/errorBoundary';
-import loader from '../../ignitus-Assets/Images/loader2.gif';
+import {withErrorBoundary} from '../../../ignitus-Shared/Components/errorBoundary';
+import loader from '../../../ignitus-Assets/Images/loader2.gif';
 import {TeamPropType, GitHubDataType} from '../types';
 import * as S from '../Styles';
