+++ src/ignitus-AboutPage/ignitus-Team/sagas.ts
@@ -1,6 +1,6 @@
 import { effects } from 'redux-saga';
 import * as t from './actionTypes';
-import * as api from '../ignitus-Api';;
+import * as api from '../../ignitus-Api';;
 
 const {
   call, put, takeLatest, all,
