+++ src/rootReducer.ts
@@ -2,7 +2,7 @@ import { combineReducers } from 'redux';
 import testimonial from './ignitus-HomePage/ignitus-Testimonial';
 import studentSignUp from './ignitus-Authentication/ignitus-StudentSignUp';
 import studentLogin from './ignitus-Authentication/ignitus-StudentLogin';
-import contributors from './ignitus-Team';
+import contributors from './ignitus-AboutPage/ignitus-Team';
 
 
 const rootReducer = combineReducers({
