+++ src/rootSagas.ts
@@ -3,7 +3,7 @@ import { all } from 'redux-saga/effects';
 import Testimonial from './ignitus-HomePage/ignitus-Testimonial';
 import studentSignUp from './ignitus-Authentication/ignitus-StudentSignUp';
 import studentLogin from './ignitus-Authentication/ignitus-StudentLogin';
-import contributors from './ignitus-Team';
+import contributors from './ignitus-AboutPage/ignitus-Team';
 
 export default function* rootSaga() {
   yield all([
